% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{coef_boot}
\alias{coef_boot}
\title{Percentile Bootstrap Estimation of canonical correlation coefficients}
\usage{
coef_boot(
  X_FIT,
  Y_FIT,
  ncomp = 10,
  Nboot = 30,
  ProcrustX = NULL,
  ProcrustY = NULL
)
}
\arguments{
\item{X_FIT}{Numeric Matrix [N, P1] containing the predictor variables.}

\item{Y_FIT}{Numeric Matrix [N, P2] containing the outcome variables.}

\item{ncomp}{Numeric Scalar. Number of CCA components to keep in analyses. Must be equal to or less than min(P1,P2).}

\item{Nboot}{Numeric Scaler. Number of times to repeat bootstrap resampling.}

\item{ProcrustX}{Numeric Matrix [ncomp, P2] containing target matrix for Procrustes Analysis. All CCA outcome raw coefficients obtained during the bootstrap resampling will be rotated to this target matrix.}
}
\value{

}
\description{
Percentile Bootstrap Estimation of canonical correlation coefficients
}
