% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{.cca}
\alias{.cca}
\title{Canonical Correlations}
\usage{
.cca(
  X_FIT,
  Y_FIT,
  X_PRED = NULL,
  Y_PRED = NULL,
  ncomp = 10,
  ProcrustX = NULL,
  ProcrustY = NULL,
  SafetyChecks = TRUE
)
}
\arguments{
\item{X_FIT}{Numeric Matrix [N, P1] containing the training dataset predictor variables.}

\item{Y_FIT}{Numeric Matrix [N, P2] containing the training dataset outcome variables.}

\item{X_PRED}{Numeric Matrix [N, P1] containing the testing dataset predictor variables. Variables should be ordered in the same way as for X_FIT.}

\item{Y_PRED}{Numeric Matrix [N, P2] containing the testing dataset outcome variables. Variables should be ordered in the same way as for Y_FIT.}

\item{ncomp}{Numeric Scalar. Number of CCA components to keep in analyses. Must be equal to or less than min(P1,P2).}

\item{ProcrustX}{Numeric Matrix [ncomp, P1] containing target matrix for Procrustes Analysis. Will align raw coefficient matrix to ProcrustX target matrix.}

\item{ProcrustY}{Numeric Matrix [ncomp, P2] containing target matrix for Procrustes Analysis. Will align raw coefficient matrix to ProcrustY target matrix.}

\item{SafetyChecks}{Checks the input provided for mistakes (default = FALSE).}
}
\value{
A list containing the following components
\itemize{
  \item xcoef - Estimated raw coefficients (CCA weights) for the x (predictor) variables.
  \item ycoef - Estimated raw coefficients (CCA weights) for the y (outcome) variables.
  \item variates - Variates (latent variable scores), estimated from the raw coefficient in X_FIT/Y_FIT.
  cbind.data.frame(xvariates, yvariates). If X_PRED and Y_PRED are provided, then  variates (and xvariates/yvariates) will be the predicted latent variable scores  from X_PRED/Y_PRED matrices.
  \item xvariates - Variates (latent variable scores) for predictor variables.
  \item yvariates - Variates (latent variable scores) for outcome variables.
  \item cc_pred - Predicted Canonical Correlations, estimated using from X_PRED & Y_PRED if provided. If no PRED matrices are specified, then this simply returns the estimated canonical correlations from the training datsets (X_FIT/Y_FIT).
  \item cc_fit - Estimated Canonical Correlation estimated from the training datasets (X_FIT/Y_FIT).
}
}
\description{
Sightly quicker than stats::cancor, and allows you to easily fit cca models in one dataset and find predicted variates/canonical correlations in another dataset.

The number of predictor (X) and outcome (Y) variables are denoted by P1 and P2 below, and N is sample size.

The function also allows the user to rotate the raw coefficients using Procrustes Analyses to target matrices (ProcrustX & ProcrustY), prior to estimating variates.
}
