% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plotcor}
\alias{plotcor}
\title{Create a ggplot2 Correlation Matrix}
\usage{
plotcor(
  dat,
  Variables_Labels = NULL,
  textadjust = 2,
  includeN = TRUE,
  reportCI = TRUE,
  low_colour = "#EDCB64",
  high_colour = "#B62A3D",
  abs_colour = TRUE,
  cluster_variables = FALSE
)
}
\arguments{
\item{dat}{Input dataframe or matrix (*do not input a correlation matrix*).}

\item{Variables_Labels}{Character vector of variable Labels, corresponding to each column in dat. If missing (NULL) then colnames(dat) will be used.}

\item{textadjust}{Scalar. Adjust text size by a magnification factor.}

\item{includeN}{Logical. Include sample size on upper diagononal (TRUE) or leave blank (FALSE).}

\item{reportCI}{Logical. Include confidence interval on upper diagonal (TRUE) or leave blank (FALSE).}

\item{low_colour}{Logical. Hex colour code for the lowest correlation.}

\item{high_colour}{Logical. Hex colour code for the highest correlation.}

\item{abs_colour}{Logical. If TRUE, will use the absolute correlation (i.e. ignoring whether the correlation is positive or negative) for determining square colour.}
}
\value{
ggplot
}
\description{
Create a ggplot2 Correlation Matrix
}
\examples{
X = sapply(1:10, function(i) rnorm(100))
X = as.data.frame(X)
My_Labels = c(paste0("Predictor ",1:5), paste0("Outcome ",1:5))

plotcor(X, Variables_Labels = My_Labels, includeN = TRUE, reportCI = FALSE)

}
