% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{cca_splithalf}
\alias{cca_splithalf}
\title{Split-Half CCA code}
\usage{
cca_splithalf(X_FIT, Y_FIT, X_PRED, Y_PRED, ncomp = NULL, alpha = 0.05)
}
\arguments{
\item{X_FIT}{Numeric Matrix [N, P1] containing the training dataset predictor variables.}

\item{Y_FIT}{Numeric Matrix [N, P2] containing the training dataset outcome variables.}

\item{X_PRED}{Numeric Matrix [N, P1] containing the testing dataset predictor variables. Variables should be ordered in the same way as for X_FIT.}

\item{Y_PRED}{Numeric Matrix [N, P1] containing the testing dataset outcome variables. Variables should be ordered in the same way as for Y_FIT.}

\item{ncomp}{Numeric Scalar. Number of CCA components to keep in analyses. Must be equal to or less than min(P1,P2).}

\item{alpha}{Numeric Scalar. Alpha level for estimating a 100(1-alpha)\% confidence interval for each canonical correlation. Default is .05 for estimating a 95\% confidence interval.}
}
\description{
Run CCA model in training dataset, and and validate performance in testing dataset.
The function estimates confidence intervals and p-values using the ... .
}
